Completely changed my code around so that not everything happens in function onLocationFound(e)
I unfortunately forgot to make commits during that process, but here is what I did;
Essentially, everything in regards to my markers was happening inside onLocationFound(e), so I made individual functions. However, detecting the current user's location had to happen inside it, as it is how the library works, so to use it outside, I had to find a way to instantiate the pin out of it while still using user's current location. This was possible by passing parameters around.

Also, I have added two buttons: One for new pin, one for relocate user.
To add a pin, you now have to first click on new, and then you click on the map to choose a a location, and then when you click the "+" the pin is immobile and it will be added at that location.